python-cli/pyproject.toml
[project]
name = "convex-notes"
version = "0.1.0"
description = "Convex Notes CLI - A beautiful terminal notes manager"
readme = "README.md"
authors = [
    { name = "Sabry", email = "dr.sabry1997@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "convex>=0.7.0",
    "typer>=0.21.0",
    "rich>=14.2.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
convex-notes = "convex_notes.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/convex_notes"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
